+++ src/app/report/[id]/_components/Card.tsx
